% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_Rttest.R
\name{new_Rttest}
\alias{new_Rttest}
\title{Title}
\usage{
new_Rttest(x, y, alpha)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{alpha}{a parameter of t.test}
}
\value{
an object of Rttest class
}
\description{
Title
}
\examples{
\dontrun{obj <- new_Rttest(x,y, 0.05)}

}
