% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Title}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{an object of class Rttest}

\item{...}{other parameters}
}
\value{
prints the object
}
\description{
Title
}
